Split out Provider and Context into separate files
